cleanup and skelton data processing

+++ BluetoothA2DPSink.h
@@ -238,7 +238,7 @@ class BluetoothA2DPSink : public BluetoothA2DPCommon {
   virtual void set_raw_stream_reader(void (*callBack)(const uint8_t *,
                                                       uint32_t));
 
-  virtual void set_raw_stream_reader_writer(void (*callBack)(uint8_t *,
+  virtual void set_raw_stream_reader_writer(void (*callBack)(int16_t *,
                                                              uint32_t));
 
   /// Define callback which is called when we receive data
@@ -408,7 +408,7 @@ class BluetoothA2DPSink : public BluetoothA2DPCommon {
   void (*data_received)() = nullptr;
   void (*stream_reader)(const uint8_t *, uint32_t) = nullptr;
   void (*raw_stream_reader)(const uint8_t *, uint32_t) = nullptr;
-  void (*raw_stream_reader_writer)(uint8_t *, uint32_t) = nullptr;
+  void (*raw_stream_reader_writer)(int16_t *, uint32_t) = nullptr;
   void (*avrc_connection_state_callback)(bool connected) = nullptr;
   void (*avrc_metadata_callback)(uint8_t, const uint8_t *) = nullptr;
 #if ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(4, 0, 0)
